moved NullSerializer and builder to satviz.serial
