+++ satviz-network/src/main/java/edu/kit/satviz/network/MessageTypes.java
@@ -1,9 +1,7 @@
 package edu.kit.satviz.network;
 
-import edu.kit.satviz.serial.ClauseSerializer;
-import edu.kit.satviz.serial.SatAssignmentSerializer;
-import edu.kit.satviz.serial.StringMapSerializer;
-import edu.kit.satviz.serial.StringSerializer;
+import edu.kit.satviz.serial.*;
+
 import java.util.Map;
 
 /**
